fix(server): avoid self-imports in normalized dts entrypoints
