+++ packages/server/tsdown.config.ts
@@ -1,5 +1,6 @@
+import { readFile, writeFile } from 'node:fs/promises';
+import path from 'node:path';
 import { defineConfig } from 'tsdown';
-import { normalizeServerDeclarationChunkImports } from './src/internals/normalizeDeclarationChunkImports.ts';
 
 export const input = [
   'src/adapters/aws-lambda/index.ts',
@@ -18,6 +19,49 @@ export const input = [
   'src/shared.ts',
   'src/unstable-core-do-not-import.ts',
 ];
+
+const UNSTABLE_CORE_ENTRYPOINT = 'src/unstable-core-do-not-import.ts';
+
+const UNSTABLE_CORE_CHUNK_IMPORT =
+  /((?:\.\.\/|\.\/)+unstable-core-do-not-import)\.d-[A-Za-z0-9_-]+\.(mjs|cjs)/g;
+
+function normalizeDeclarationChunkImports(contents: string) {
+  return contents.replace(
+    UNSTABLE_CORE_CHUNK_IMPORT,
+    (_, importPath: string, format: 'mjs' | 'cjs') =>
+      `${importPath}.d.${format === 'mjs' ? 'mts' : 'cts'}`,
+  );
+}
+
+function getDeclarationOutputs(rawInputs: string[]) {
+  return rawInputs
+    .filter((entrypoint) => entrypoint !== UNSTABLE_CORE_ENTRYPOINT)
+    .flatMap((entrypoint) => {
+      const distBase = entrypoint
+        .split('/')
+        .slice(1)
+        .join('/')
+        .replace(/\.(ts|tsx)$/, '');
+
+      return [`${distBase}.d.mts`, `${distBase}.d.cts`];
+    });
+}
+
+async function normalizeServerDeclarationChunkImports(rawInputs: string[]) {
+  const declarationOutputs = getDeclarationOutputs(rawInputs);
+
+  await Promise.all(
+    declarationOutputs.map(async (relativePath) => {
+      const filePath = path.resolve('dist', relativePath);
+      const original = await readFile(filePath, 'utf8');
+      const normalized = normalizeDeclarationChunkImports(original);
+
+      if (normalized !== original) {
+        await writeFile(filePath, normalized, 'utf8');
+      }
+    }),
+  );
+}
 
 export default defineConfig({
   target: ['node18', 'es2017'],
